requirements.txt
discord.py>=2.3.2
wavelink>=2.5.0
python-dotenv
